Minor code style changes
