+++ Pitter/src/api/__init__.py
@@ -1,0 +1,4 @@
+from __future__ import absolute_import, unicode_literals
+from .celery import app as celery_app
+
+__all__=('celery_app',)
