+++ Pitter/src/api/utils/data/text_tools.py
@@ -1,0 +1,9 @@
+POSITIVE_CASES = dict(
+    kode='edok',
+    python='nohtyp'
+)
+NEGATIVE_CASES = dict(
+    ''=''
+
+
+)
