+++ Pitter/src/api/views/__init__.py
@@ -1,5 +1,10 @@
 from .users import CreateUserAPIView
-
+from .subscribers import GetFollowersList, FollowUser
+from .search import UserSearch
 __all__ = [
-    'CreateUserAPIView'
+    'CreateUserAPIView',
+    'GetFollowersList',
+    'FollowUser',
+    'UserSearch'
+
 ]
