+++ Pitter/src/api/views/obtaintoken.py
@@ -1,8 +1,7 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.permissions import AllowAny
-from models.user_models import User
-import jwt
+from ..models.user_models import User
 from rest_framework_jwt.settings import api_settings
