+++ Pitter/src/api/views/search.py
@@ -1,11 +1,12 @@
 from rest_framework.permissions import IsAuthenticated, AllowAny
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from models.user_models import User
-from models.user_serializers import UserSerializer
+from ..models.user_models import User
+from ..models.user_serializers import UserSerializer
 
 
 class UserSearch(APIView):
+    # todo AUTHORIZE
     permission_classes = (AllowAny,)
 
     def get(self, request, person):
@@ -17,7 +18,7 @@ class UserSearch(APIView):
         try:
             user = User.objects.get(username=person)
         except User.DoesNotExist:
-            return Response({'message': 'Not found.'}, status= 400)
+            return Response({'message': 'Not found.'}, status=400)
 
         serializer = UserSerializer(user)
