+++ Pitter/src/requirements.txt
@@ -6,4 +6,4 @@ redis==3.2.1
 django_redis==4.10.0
 requests==2.21.0
 djangorestframework-jwt==1.11.0
-celery==4.3.0
+celery>=4.2,<4.3
